requirements.txt
aiohttp
discord.py
python-dotenv
beautifulsoup4